Refactor/reorganize CampaignList and related views + NavigationStack
