--- Shared/Views/App/CampaignList.swift
@@ -1,859 +1,0 @@
-//
-//  CampaignList.swift
-//  St Jude
-//
-//  Created by Ben Cardy on 18/08/2022.
-//
-
-import SwiftUI
-import GRDB
-
-enum FundraiserSortOrder: Int, CaseIterable {
-    case byName
-    case byAmountRaised
-    case byAmountRemaining
-    case byGoal
-    case byPercentage
-    
-    var description: String {
-        switch self {
-        case .byName:
-            return "Name"
-        case .byAmountRaised:
-            return "Amount Raised"
-        case .byGoal:
-            return "Goal"
-        case .byPercentage:
-            return "Percentage"
-        case .byAmountRemaining:
-            return "Amount Remaining"
-        }
-    }
-    
-    var iconName: String {
-        switch self {
-        case .byName:
-            return "characters.lowercase"
-        case .byAmountRaised, .byAmountRemaining:
-            return "dollarsign.arrow.trianglehead.counterclockwise.rotate.90"
-        case .byGoal:
-            return "dollarsign.circle.fill"
-        case .byPercentage:
-            return "percent"
-        }
-    }
-}
-
-enum CampaignListSheet: Identifiable {
-    case aboutScreen
-    case leaderBoard
-    case randomPicker
-    case easterEgg
-    case startHeadToHead
-    case continueHeadToHead(campaign: Campaign)
-    
-    var id: String {
-        switch self {
-        case .aboutScreen:
-            return "aboutScreen"
-        case .leaderBoard:
-            return "leaderBoard"
-        case .randomPicker:
-            return "randomPicker"
-        case .easterEgg:
-            return "easterEgg"
-        case .startHeadToHead:
-            return "startHeadToHead"
-        case let .continueHeadToHead(campaign):
-            return "continueHeadToHead:\(campaign.id.uuidString)"
-        }
-    }
-    
-}
-
-struct CampaignList: View {
-
-    init() {
-        updateNavBarFont()
-    }
-
-    func updateNavBarFont() {
-//        UINavigationBar.appearance().titleTextAttributes = [.font : UserDefaults.shared.disablePixelFont ? UIFont.preferredFont(forTextStyle: .headline) : UIFont(name: Font.customFontName, size: UIFont.preferredFont(forTextStyle: .headline).pointSize) ?? UIFont.systemFont(ofSize: 20)]
-//        UINavigationBar.appearance().largeTitleTextAttributes = [.font : UserDefaults.shared.disablePixelFont ? UIFont.preferredFont(forTextStyle: .largeTitle) : UIFont(name: Font.customFontName, size: UIFont.preferredFont(forTextStyle: .largeTitle).pointSize)  ?? UIFont.systemFont(ofSize: 20)]
-    }
-    
-    @Environment(\.colorScheme) var colorScheme
-    @Environment(\.dynamicTypeSize) var dynamicTypeSize
-    
-    // MARK: 2025
-    private func getNewCompetitors() {
-        // Ensure we don't select either of the current players
-        let newCompetitor1 = Player.allCases.filter {
-            $0 != self.competitor1 && $0 != self.competitor2
-        }.randomElement() ?? .myke
-        let newCompetitor2 = Player.allCases.filter {
-            $0 != self.competitor1 && $0 != self.competitor2 && $0 != newCompetitor1
-        }.randomElement() ?? .stephen
-        
-        self.competitor1 = newCompetitor1
-        self.competitor2 = newCompetitor2
-    }
-    
-    @State private var competitor1: Player = .myke
-    @State private var competitor2: Player = .stephen
-    
-    // MARK: 2023
-    @State private var teamEvent: TeamEvent? = nil
-    @State private var teamEventObservation = AppDatabase.shared.observeTeamEventObservation()
-    @State private var teamEventCancellable: DatabaseCancellable?
-    
-    @State private var showSheet: CampaignListSheet? = nil
-    
-    // MARK: 2022
-    
-    @State private var allCampaigns: [Campaign] = []
-    @State private var headToHeads: [HeadToHeadWithCampaigns] = []
-    
-    @State private var fundraiserSortOrder: FundraiserSortOrder = .byName
-    @State private var compactListMode: Bool = false
-    @State private var selectedCampaignId: UUID? = nil
-    
-    @State private var showSearchBar: Bool = false
-    @State private var searchText = ""
-    
-    @State private var isRefreshing: Bool = false
-    @State private var isLoading: Bool = true
-    @State private var showStephen: Bool = false
-    
-    @State private var showHeadToHeads: Bool = true
-    @State private var rotationAnimation: Bool = false
-    
-    @AppStorage(UserDefaults.iconsUnlockedKey, store: UserDefaults.shared) private var iconsUnlocked: Bool = false
-    
-    let timer = Timer.publish(every: 60, on: .main, in: .common).autoconnect()
-    
-    func compareNames(c1: Campaign, c2: Campaign) -> Bool {
-        if c1.name.lowercased() == c2.name.lowercased() {
-            return c1.id.uuidString < c2.id.uuidString
-        }
-        return c1.name.lowercased() < c2.name.lowercased()
-    }
-    
-    func sortCampaigns(_ campaigns: [Campaign]) -> [Campaign] {
-        return campaigns.sorted { c1, c2 in
-            if c1.isStarred && !c2.isStarred {
-                return true
-            }
-            if c2.isStarred && !c1.isStarred {
-                return false
-            }
-            switch fundraiserSortOrder {
-            case .byAmountRaised:
-                let v1 = c1.totalRaisedNumerical
-                let v2 = c2.totalRaisedNumerical
-                if v1 == v2 {
-                    return compareNames(c1: c1, c2: c2)
-                }
-                return v1 > v2
-            case .byGoal:
-                let v1 = c1.goalNumerical
-                let v2 = c2.goalNumerical
-                if v1 == v2 {
-                    return compareNames(c1: c1, c2: c2)
-                }
-                return v1 > v2
-            case .byPercentage:
-                let v1 = c1.percentageReached ?? 0
-                let v2 = c2.percentageReached ?? 0
-                if v1 == v2 {
-                    return compareNames(c1: c1, c2: c2)
-                }
-                return v1 > v2
-            case .byAmountRemaining:
-                var v1 = c1.goalNumerical - c1.totalRaisedNumerical
-                var v2 = c2.goalNumerical - c2.totalRaisedNumerical
-                if v1 <= 0 {
-                    v1 = .infinity
-                }
-                if v2 <= 0 {
-                    v2 = .infinity
-                }
-                if v1 == v2 {
-                    return compareNames(c1: c1, c2: c2)
-                }
-                return v1 < v2
-            default:
-                return compareNames(c1: c1, c2: c2)
-            }
-        }
-    }
-    
-    var searchResults: [Campaign] {
-        let query = searchText.lowercased().trimmingCharacters(in: .whitespacesAndNewlines)
-        if query.isEmpty {
-            return allCampaigns
-        } else {
-            return allCampaigns.filter { $0.title.lowercased().contains(query) || $0.user.username.lowercased().contains(query) }
-        }
-    }
-    
-    @ViewBuilder
-    var headToHeadList: some View {
-        ForEach(headToHeads, id: \.headToHead.id) { headToHead in
-            NavigationLink(destination: HeadToHeadView(campaign1: headToHead.campaign1, campaign2: headToHead.campaign2).tint(.white)) {
-                HeadToHeadListItem(headToHead: headToHead)
-            }
-            .contextMenu {
-                Button(role: .destructive) {
-                    Task {
-                        do {
-                            try await AppDatabase.shared.deleteHeadToHead(headToHead.headToHead)
-                        } catch {
-                            dataLogger.error("Could not delete head to head: \(error.localizedDescription)")
-                        }
-                        await fetch()
-                    }
-                } label: {
-                    Label("Remove Head to Head", systemImage: "trash")
-                }
-            }
-        }
-        .compositingGroup()
-    }
-    
-    @ViewBuilder
-    var topView: some View {
-        VStack(spacing: 0) {
-            Group {
-                VStack {
-                    Group {
-                        if let teamEvent = teamEvent {
-                            NavigationLink(destination: CampaignView(teamEvent: teamEvent), tag: teamEvent.id, selection: $selectedCampaignId) {
-                                TeamEventCardView(teamEvent: teamEvent, showDisclosureIndicator: true, showShareSheet: .constant(false), showBackground: false)
-                            }
-                            .buttonStyle(PrimaryButtonStyle(useCapsuleShape: false, useBoldText: false))
-                            .padding()
-                        } else {
-                            TeamEventCardView(teamEvent: teamEvent, showDisclosureIndicator: true, showShareSheet: .constant(false))
-                                .padding()
-                                .foregroundStyle(Color.contentColorForAccent)
-                        }
-                    }
-                }
-            }
-            .frame(maxWidth: Double.stretchedContentMaxWidth)
-            
-        }
-        .frame(maxWidth: .infinity)
-    }
-        
-    @ViewBuilder
-    var headToHeadListView: some View {
-        GroupBox {
-            VStack {
-                Button(action: {
-                    withAnimation {
-                        showHeadToHeads.toggle()
-                    }
-                }) {
-                    HStack {
-                        Text("Head to Head")
-                            .font(.title2)
-                            .fontWeight(.bold)
-                        if headToHeads.count > 0 {
-                            Button(action: {
-                                showSheet = .startHeadToHead
-                            }) {
-                                Label("Start Head to Head", systemImage: "plus").labelStyle(.iconOnly)
-                                    .fontWeight(.bold)
-                            }
-                            .foregroundStyle(Color.accentColor)
-                            .padding(.horizontal, 4)
-                            .padding(.vertical, 4)
-                            .aspectRatio(1.0, contentMode: .fit)
-                        }
-                        Spacer()
-                        Image(systemName: "chevron.right")
-                            .foregroundStyle(.secondary)
-                            .rotationEffect(.degrees(showHeadToHeads ? 90 : 0))
-                    }
-                    .contentShape(Rectangle())
-                }
-                .buttonStyle(.plain)
-                if headToHeads.count == 0 {
-                    if showHeadToHeads {
-                        VStack {
-                            Button(action: {
-                                showSheet = .startHeadToHead
-                            }, label: {
-                                Text("Add a Head to Head")
-                                    .font(.headline)
-                                    .frame(minWidth: 0, maxWidth: .infinity, alignment: .center)
-                            })
-                            .buttonStyle(PrimaryButtonStyle())
-                        }
-                    }
-                } else {
-                    if showHeadToHeads {
-                        LazyVGrid(columns: [GridItem(.adaptive(minimum: 300, maximum: .infinity), alignment: .top)], spacing: 10) {
-                            headToHeadList
-                        }
-                    }
-                }
-            }
-        }
-        .padding(.horizontal)
-        .frame(maxWidth: Double.stretchedContentMaxWidth)
-    }
-    
-    @ViewBuilder
-    func fundraiserHeaderView(scrollViewReader: SwiftUI.ScrollViewProxy) -> some View {
-        Group {
-            if allCampaigns.count != 0 {
-                if dynamicTypeSize >= .xLarge {
-                    GroupBox {
-                        VStack(spacing: 0) {
-                            HStack {
-                                Text("Fundraisers")
-                                    .font(.title2)
-                                    .fontWeight(.bold)
-                                Spacer()
-                            }
-                            HStack {
-                                Text("\(allCampaigns.count)")
-                                    .foregroundColor(.secondary)
-                                    .background(
-                                        Color.secondarySystemBackground
-                                            .cornerRadius(15)
-                                    )
-                                Spacer()
-                                Button(action: {
-                                    showSheet = .leaderBoard
-                                }) {
-                                    Label("Leaderboard", systemImage: "trophy")
-                                        .labelStyle(.iconOnly)
-                                }
-                                Menu {
-                                    ForEach(FundraiserSortOrder.allCases, id: \.self) { order in
-                                        Button(action: {
-                                            withAnimation {
-                                                fundraiserSortOrder = order
-                                            }
-                                        }) {
-                                            Label("Sort by \(order.description)", systemImage: fundraiserSortOrder == order ? "checkmark" : "")
-                                        }
-                                    }
-                                    Divider()
-                                    Button(action: {
-                                        withAnimation {
-                                            compactListMode.toggle()
-                                        }
-                                    }) {
-                                        Label("Compact View", systemImage: compactListMode ? "checkmark" : "rectangle.compress.vertical")
-                                    }
-                                } label: {
-                                    Image(systemName: "gear")
-                                }
-                                Button(action: {
-                                    withAnimation {
-                                        showSearchBar = true
-                                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                                            withAnimation {
-                                                scrollViewReader.scrollTo("SEARCH_BAR", anchor: .top)
-                                            }
-                                        }
-                                    }
-                                }) {
-                                    Label("Search", systemImage: "magnifyingglass")
-                                        .labelStyle(.iconOnly)
-                                }
-                            }
-                        }
-                    }
-                    .padding(.horizontal)
-                } else {
-                    GroupBox {
-                        HStack {
-                            Text("Fundraisers")
-                                .font(.title2)
-                                .fontWeight(.bold)
-                            Spacer()
-                            Text("\(allCampaigns.count)")
-                                .foregroundColor(.secondary)
-                                .padding(.horizontal, 8)
-                                .padding(.vertical, 4)
-                                .background(
-                                    Color.secondarySystemBackground
-                                        .cornerRadius(15)
-                                )
-                            Button(action: {
-                                showSheet = .leaderBoard
-                            }) {
-                                Label("Leaderboard", systemImage: "trophy")
-                                    .labelStyle(.iconOnly)
-                            }
-                            Menu {
-                                ForEach(FundraiserSortOrder.allCases, id: \.rawValue) { order in
-                                    Button(action: {
-                                        withAnimation {
-                                            fundraiserSortOrder = order
-                                        }
-                                    }) {
-                                        Label("Sort by \(order.description)", systemImage: fundraiserSortOrder == order ? "checkmark" : "")
-                                    }
-                                }
-                                Divider()
-                                Button(action: {
-                                    withAnimation {
-                                        compactListMode.toggle()
-                                    }
-                                }) {
-                                    Label("Compact View", systemImage: compactListMode ? "checkmark" : "rectangle.compress.vertical")
-                                }
-                            } label: {
-                                Image(systemName: "gear")
-                            }
-                            Button(action: {
-                                withAnimation {
-                                    showSearchBar = true
-                                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                                        withAnimation {
-                                            scrollViewReader.scrollTo("SEARCH_BAR", anchor: .top)
-                                        }
-                                    }
-                                }
-                            }) {
-                                Label("Search", systemImage: "magnifyingglass")
-                                    .labelStyle(.iconOnly)
-                            }
-                        }
-                    }
-                    .padding(.horizontal)
-                }
-            }
-
-            if allCampaigns.count != 0 {
-                
-                if showSearchBar {
-                    GroupBox {
-                        SearchBar(text: $searchText, placeholder: "Search...", showingMyself: $showSearchBar)
-                            .id("SEARCH_BAR")
-                    }
-                    .padding(.horizontal)
-                }
-                Button(action: {
-                    showSheet = .randomPicker
-                }) {
-                    HStack {
-                        Text("Spin for a random Fundraiser!")
-                            .fontWeight(.bold)
-                            .multilineTextAlignment(.leading)
-                            .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-                        Image(systemName: "chevron.right")
-                    }
-                }
-                .padding(.horizontal)
-                .buttonStyle(PrimaryButtonStyle())
-            } else {
-                Group {
-                    if isLoading {
-                        GroupBox {
-                            VStack {
-                                ProgressView()
-                                    .padding(.top, 40)
-                                    .padding(.bottom, 10)
-                                Text("Loading ...")
-                                    .padding(.bottom, 40)
-                            }
-                            .frame(maxWidth: .infinity)
-                        }
-                        
-                    } else {
-                        GroupBox {
-                            VStack {
-                                Image(systemName: "exclamationmark.triangle")
-                                    .padding(.top, 40)
-                                    .padding(.bottom, 10)
-                                Text("No fundraisers yet")
-                                Link(destination: URL(string: "https://start.tiltify.com/?supportingFactId=1c6d5c76-1804-48fa-a474-2bfe1c52f48c")!, label: {
-                                    Text("Be the first and create your own!")
-                                        .frame(minWidth: 0, maxWidth: .infinity, alignment: .center)
-                                })
-                                .fontWeight(.bold)
-                                .multilineTextAlignment(.center)
-                                .buttonStyle(PrimaryButtonStyle())
-                            }
-                            .frame(maxWidth: .infinity)
-                        }
-                    }
-                }
-                .padding(.horizontal)
-            }
-        }
-        .frame(maxWidth: Double.stretchedContentMaxWidth)
-    }
-    
-    @ViewBuilder
-    var fundraiserListView: some View {
-        if(allCampaigns.count != 0) {
-            LazyVGrid(columns: [GridItem(.adaptive(minimum: 300, maximum: .infinity), alignment: .top)]) {
-                
-                ForEach(Array(searchResults.enumerated()), id: \.offset) { index, campaign in
-                        NavigationLink(destination: CampaignView(initialCampaign: campaign)) {
-                            GroupBox {
-                                FundraiserListItem(campaign: campaign, sortOrder: fundraiserSortOrder, compact: compactListMode, showBackground: false, showShareSheet: .constant(false))
-                            }
-                        }
-                        .buttonStyle(PlainButtonStyle())
-//                        .buttonStyle(PrimaryButtonStyle(tint: .secondarySystemBackground, useCapsuleShape: false))
-                        .contextMenu {
-                            Button(action: {
-                                showSheet = .continueHeadToHead(campaign: campaign)
-                            }) {
-                                Label(title: {
-                                    Text("Start Head to Head")
-                                }, icon: {
-                                    Image(systemName: "trophy")
-                                })
-                            }
-                            Button(action: {
-                                Task {
-                                    await starOrUnstar(campaign: campaign)
-                                }
-                            }) {
-                                
-                                Label(title: {
-                                    Text(campaign.isStarred ? "Unfavourite" : "Favourite")
-                                }, icon: {
-                                    Image(systemName: campaign.isStarred ? "heart.fill" : "heart")
-                                })
-                            }
-                        }
-                }
-            }
-            .padding(.horizontal)
-            // TODO: add these back
-//            if searchText.lowercased() == "jonycube" || searchText.lowercased() == "jony cube" {
-//                Image.imageAtScale(.jonycubePixel2024, scale: 0.5)
-//            } else if searchText.lowercased() == "l2cu" {
-//                Image.imageAtScale(.l2CuPixel2024, scale: 0.5)
-//            }
-        }
-    }
-    
-    @ViewBuilder
-    var easterEggView: some View {
-        Button(action: {
-            showSheet = .easterEgg
-        }, label: {
-            HStack {
-                Text("App from the Lovely Developers")
-                    .font(.caption)
-                    .bold()
-                Image(.l2CuHeadOutline)
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(width: 30, height: 30)
-                    .accessibility(hidden: true)
-                
-            }
-            .frame(maxWidth: .infinity)
-        })
-        .buttonStyle(PrimaryButtonStyle())
-        .foregroundStyle(.black)
-        .padding(.horizontal)
-        .frame(maxWidth: Double.stretchedContentMaxWidth)
-    }
-    
-    @ViewBuilder
-    var widgetCompatibilityView: some View {
-        //                        if selectedCampaignId != nil {
-        /// In order to open a selected campaign when a widget is tapped, the corresponding
-        /// NavigationLink needs to be loaded. That  isn't guaranteed when they are presented
-        /// in a Lazy grid as below, so we create a bunch of empty/invisible NavigationLinks to
-        /// trigger on the widget tap instead
-        ForEach(allCampaigns, id: \.id) { campaign in
-            NavigationLink(destination: CampaignView(initialCampaign: campaign), tag: campaign.id, selection: $selectedCampaignId) {
-                EmptyView()
-            }
-        }
-        
-        ForEach(headToHeads, id: \.headToHead.id) { headToHead in
-            NavigationLink(destination: HeadToHeadView(campaign1: headToHead.campaign1, campaign2: headToHead.campaign2), tag: headToHead.headToHead.id, selection: $selectedCampaignId) {
-                EmptyView()
-            }
-        }
-    }
-    
-    var body: some View {
-        ScrollView {
-            ScrollViewReader { scrollViewReader in
-                VStack(spacing: 0) {
-                    VStack{
-                        topView
-                    }
-
-                    VStack {
-                        CountdownView()
-                            .padding()
-                        headToHeadListView
-                        fundraiserHeaderView(scrollViewReader: scrollViewReader)
-                        fundraiserListView
-                        easterEggView
-                        widgetCompatibilityView
-                    }
-                    .padding(.top,-10)
-                    .frame(maxWidth: .infinity)
-                }
-                .rotationEffect(Angle(degrees: rotationAnimation ? 0 : 360))
-                
-            }
-        }
-        .refreshable {
-            await refresh()
-        }
-        .onReceive(timer) { _ in
-            Task {
-                await refresh()
-            }
-        }
-        .onChange(of: fundraiserSortOrder) { newValue in
-            UserDefaults.shared.campaignListSortOrder = newValue
-            Task {
-                let newAllCampaigns = sortCampaigns(allCampaigns)
-                DispatchQueue.main.async {
-                    allCampaigns = newAllCampaigns
-                }
-            }
-        }
-        .onChange(of: compactListMode) { newValue in
-            UserDefaults.shared.campaignListCompactView = newValue
-        }
-        .onChange(of: showHeadToHeads) { newValue in
-            UserDefaults.shared.expandHeadToHeadSection = newValue
-        }
-        .onAppear {
-            showStephen = Bool.random()
-            showHeadToHeads = UserDefaults.shared.expandHeadToHeadSection
-            fundraiserSortOrder = UserDefaults.shared.campaignListSortOrder
-            compactListMode = UserDefaults.shared.campaignListCompactView
-            
-            teamEventCancellable = AppDatabase.shared.start(observation: teamEventObservation) { error in
-                dataLogger.error("Error observing stored team event: \(error.localizedDescription)")
-            } onChange: { event in
-                teamEvent = event
-            }
-            
-            Task {
-                await fetch()
-            }
-            Task {
-                await refresh()
-            }
-            
-        }
-        .sheet(item: $showSheet, onDismiss: {
-            SoundEffectHelper.shared.stop()
-        }) { sheet in
-            switch sheet {
-            case .aboutScreen:
-                AboutView()
-                    .forSheet()
-            case .leaderBoard:
-                Leaderboard(campaigns: allCampaigns) { campaign in
-                    showSheet = nil
-                    selectedCampaignId = campaign.id
-                }
-                .forSheet(displayMode: .large)
-            case .randomPicker:
-                    RandomCampaignPickerView(campaignChoiceID: self.$selectedCampaignId, allCampaigns: allCampaigns)
-                        .forSheet()
-            case .easterEgg:
-                EasterEggView()
-                    .forSheet(displayMode: .large)
-            case .startHeadToHead:
-                ChooseCampaignView() { campaign in
-                    showSheet = .continueHeadToHead(campaign: campaign)
-                }
-                .forSheet(displayMode: .large)
-            case let .continueHeadToHead(firstCampaign):
-                ChooseCampaignView(otherCampaign: firstCampaign) { otherCampaign in
-                    Task {
-                        let headToHead = HeadToHead(id: UUID(), campaignId1: firstCampaign.id, campaignId2: otherCampaign.id)
-                        do {
-                            try await AppDatabase.shared.saveHeadToHead(headToHead)
-                        } catch {
-                            dataLogger.error("Could not create Head to Head: \(error.localizedDescription)")
-                        }
-                        await fetch()
-                        selectedCampaignId = headToHead.id
-                    }
-                }
-                .forSheet(displayMode: .large)
-            }
-        }
-        .toolbar {
-            ToolbarItem(placement: .navigationBarLeading) {
-                Button(action: {
-                    showSheet = .aboutScreen
-                }) {
-                    Image(systemName: "info")
-                }
-            }
-            
-            ToolbarItem(placement: .navigationBarTrailing) {
-                Button(action: {
-                    isRefreshing = true
-                    Task {
-                        await refresh()
-                        isRefreshing = false
-                    }
-                }) {
-                    ZStack {
-                        if isRefreshing {
-                            ProgressView()
-                        }
-                        Image(systemName: "arrow.clockwise")
-                            .opacity(isRefreshing ? 0 : 1)
-                    }
-                }
-                .keyboardShortcut("r")
-                .disabled(isRefreshing)
-            }
-        }
-        .onOpenURL { url in
-            if let components = URLComponents(url: url, resolvingAgainstBaseURL: false), components.host == "campaign", let queryComponents = components.queryItems?.reduce(into: [String: String](), { (result, item) in
-                result[item.name] = item.value
-            }), let id = queryComponents["id"] {
-                selectedCampaignId = UUID(uuidString: id)
-                showSheet = nil
-            }
-        }
-        .onChange(of: searchText, perform: { value in
-            if value.lowercased() == "do a barrel roll" {
-                rotationAnimation = false
-                withAnimation(.easeInOut(duration: 2.0)) {
-                    rotationAnimation = true
-                }
-            }
-        })
-    }
-    
-    func starOrUnstar(campaign: Campaign) async {
-        let newCampaign = campaign.setStar(to: !campaign.isStarred)
-        do {
-            if try await AppDatabase.shared.updateCampaign(newCampaign, changesFrom: campaign) {
-                dataLogger.info("Updated starring stored campaign: \(newCampaign.id)")
-            }
-        } catch {
-            dataLogger.error("Starring/unstarring stored campaign failed: \(error.localizedDescription)")
-        }
-        await fetch()
-    }
-    
-    func refresh() async {
-                        
-        if let apiEventData = await TiltifyAPIClient.shared.getFundraisingEvent() {
-            dataLogger.debug("API fetched TeamEvent: \(apiEventData.name)")
-            let apiEvent = TeamEvent(from: apiEventData)
-            do {
-                if let existingTeamEvent = teamEvent {
-                    teamEvent = apiEvent
-                    if try await AppDatabase.shared.updateTeamEvent(apiEvent, changesFrom: existingTeamEvent) {
-                        dataLogger.info("Updated team event \(apiEvent.name) (id: \(apiEvent.id)")
-                    }
-                } else {
-                    dataLogger.debug("Saved new team event")
-                    teamEvent = try! await AppDatabase.shared.saveTeamEvent(apiEvent)
-                }
-            } catch {
-                dataLogger.error("Updating stored team event failed: \(error.localizedDescription)")
-            }
-        }
-        
-        dataLogger.debug("Fetching campaigns...")
-        let apiCampaigns = await TiltifyAPIClient.shared.getFundraisingEventCampaigns()
-        var keyedApiCampaigns: [UUID: Campaign] = apiCampaigns.reduce(into: [:]) { partialResult, campaign in
-            partialResult.updateValue(Campaign(from: campaign), forKey: campaign.id)
-        }
-        dataLogger.debug("Fetching campaigns... \(apiCampaigns.count) Done")
-        
-        do {
-            // For each campaign from the database...
-            for dbCampaign in try await AppDatabase.shared.fetchAllCampaigns() {
-                if let apiCampaign = keyedApiCampaigns[dbCampaign.id] {
-                    // Update it from the API if it exists...
-                    keyedApiCampaigns.removeValue(forKey: dbCampaign.id)
-                    let updateCampaign = dbCampaign.isStarred ? apiCampaign.setStar(to: true) : apiCampaign
-                    do {
-                        dataLogger.notice("Updating \(apiCampaign.name) - \(apiCampaign.totalRaised.description(showFullCurrencySymbol: false))")
-                        try await AppDatabase.shared.updateCampaign(updateCampaign, changesFrom: dbCampaign)
-                        if apiCampaign.id == TLD_CAMPAIGN {
-                            iconsUnlocked = apiCampaign.totalRaisedNumerical >= TLDMilestones.IconsUnlocked
-                        }
-                    } catch {
-                        dataLogger.error("Failed to update campaign: \(updateCampaign.id) \(updateCampaign.name): \(error.localizedDescription)")
-                    }
-                } else {
-                    // Remove it from the database if it doesn't...
-                    do {
-                        try await AppDatabase.shared.deleteCampaign(dbCampaign)
-                    } catch {
-                        dataLogger.error("Failed to delete campaign \(dbCampaign.id) \(dbCampaign.name): \(error.localizedDescription)")
-                    }
-                }
-            }
-            // For each new campaign in the API, save it to the database
-            for apiCampaign in keyedApiCampaigns.values {
-                do {
-                    try await AppDatabase.shared.saveCampaign(apiCampaign)
-                    if apiCampaign.id == TLD_CAMPAIGN {
-                        iconsUnlocked = apiCampaign.totalRaisedNumerical >= TLDMilestones.IconsUnlocked
-                    }
-                } catch {
-                    dataLogger.error("Failed to save Campaign \(apiCampaign.id) \(apiCampaign.name): \(error.localizedDescription)")
-                }
-            }
-        } catch {
-            dataLogger.error("Could not update campaigns")
-        }
-        
-        await fetch()
-        
-        self.isLoading = false
-        
-        self.getNewCompetitors()
-        
-    }
-    
-    func fetch() async {
-        do {
-            dataLogger.notice("Fetched stored fundraiser")
-            try Task.checkCancellation()
-            allCampaigns = sortCampaigns(try await AppDatabase.shared.fetchAllCampaigns().filter { !HIDDEN_CAMPAIGN_IDS.contains($0.id) })
-            dataLogger.notice("Fetched stored campaigns")
-            let fetchedHeadToHeads = try await AppDatabase.shared.fetchAllHeadToHeads()
-            withAnimation {
-                headToHeads = fetchedHeadToHeads
-            }
-            dataLogger.notice("Fetched stored head to heads")
-            
-        } catch is CancellationError {
-            dataLogger.info("Campaign fetch cancelled")
-        }
-        catch {
-            dataLogger.error("Failed to fetch stored fundraiser: \(error.localizedDescription)")
-        }
-    }
-    
-}
-
-struct CampaignList_Previews: PreviewProvider {
-    static var previews: some View {
-        NavigationView {
-            CampaignList()
-                .navigationBarTitleDisplayMode(.inline)
-        }
-        .navigationViewStyle(.stack)
-    }
-}
