+++ Shared/Views/Legacy/NotificationSettings.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import UserNotifications
 import Combine
 
+// We'll add this back someday. 2026 maybe??
 struct NotificationSettings: View {
     @Environment(\.openURL) private var openURL
     @Environment(\.presentationMode) private var presentationMode
